adjusts how the classes interact. Also does more of the spreadsheet
We used to use the dbClass and the cluster class to hold cluster data.  This moved all cluster specific work into cluster. And this meant the other clases like ores had to be adjusted to deal with that.
There is most likely more work on the other classes to do to clear them up after this change.