db got corrupted so this attempts to fix the issue
Split the database to have 3 new tables.  Transactions, Storage and trends.  Hopefully spliting the db up like this will increase performance and decrease the chance of corruption.  It also means as long as core is still uncorrupted the other dbs can be deleted and start from the current point in time.